refactor: remove LoginRequireModal and introduce SignInModal with email and Google sign-in options
